[bug] Jitter calculation would stop
The jitter calculation would update for 12 seconds and then stop.
The reason was that after the first 60 jitter measurements, no new
ones would be added to the jitter database. Or more correctly, they
would get added, but immediately removed again :p

Now I remove the oldest measurement, instead of the one that was just added.

Fixes #1