[SignpostDemo] Updated jitter UDP protocol
Changed the UDP protocol to make it easier to code
against for non Cocoa devices:

The message format expected of the packages are:
'HOSTNAME\r\nTIMESTAMP\r\nJITTER'

The hostname is the id of the host. Same as is used
in the initial handshake.

The timestamp is the time in seconds since 1970.
The value used in the cocoa clients is a NSTimeInterval.
The following is from its documentation:

  NSTimeInterval is always specified in seconds;
  it yields sub-millisecond precision over a range
  of 10,000 years.

The jitter is also a double specifying jitter in ms.
For the sending party, the jitter should be the jitter
the client sees the in the messages received by the other
party. The server will reply with the jitter seen for
the jitter UDP probes from that particular client.