[SignpostDemo] Use concurrent queues for jitter
Jitter measurements are now executed on concurrent
queues, instead of serial queues like before.
This seems to give better performance, and anecdotal
evidence seem to show that the iOS clients also
report somewhat smaller jitter as a result as well.

The is a synchronized block that is executed in the
UDP receive block. This is a contention point in the
jitter measurement. The measurement itself is concluded
outside this block of code as the first thing upon
receiving the UDP message. Now that concurrent queues
are used, the jitter calcualtions can all be concluded
before the thread has to wait for exclusive access to
the user data variable. Gain :)